requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
rave-python==1.2.16
requests==2.28.2
SQLAlchemy==2.0.3
tqdm==4.65.0
twilio==7.16.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.2
